pyproject.toml
[tool.poetry]
name = "finn-codesubmit"
version = "0.1.0"
description = ""
authors = ["Paolo Sacconier <paolo@sacconier.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
mangum = "^0.17.0"
openapi-to-fastapi = "^0.16.0"
validators = "^0.31.0"
pybase62 = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
pytest = "^8.2.2"
skjold = "^0.6.2"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.7"

[tool.skjold]
report_only = false
sources = ["pyup", "gemnasium"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
